+++ source/Energinet.DataHub.MessageArchive.IntegrationTests/Repositories/ArchiveReaderRepositoryTests.cs
@@ -16,11 +16,11 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Energinet.DataHub.MessageArchive.Domain.Models;
-using Energinet.DataHub.MessageArchive.Domain.Repositories;
 using Energinet.DataHub.MessageArchive.EntryPoint;
 using Energinet.DataHub.MessageArchive.Persistence.Containers;
-using Energinet.DataHub.MessageArchive.Search;
+using Energinet.DataHub.MessageArchive.PersistenceModels;
+using Energinet.DataHub.MessageArchive.Reader;
+using Energinet.DataHub.MessageArchive.Reader.Models;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using SimpleInjector;
