+++ source/Energinet.DataHub.MessageArchive.PersistanceModels/BaseParsedModel.cs
@@ -15,7 +15,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace Energinet.DataHub.MessageArchive.Domain.Models
+namespace Energinet.DataHub.MessageArchive.PersistenceModels
 {
     public class BaseParsedModel
     {
