+++ source/Energinet.DataHub.MessageArchive.PersistanceModels/Energinet.DataHub.MessageArchive.PersistanceModels.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <LangVersion>9.0</LangVersion>
+    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+    <RunAnalyzersDuringBuild>true</RunAnalyzersDuringBuild>
+    <GenerateDocumentationFile>true</GenerateDocumentationFile>
+    <RootNamespace>Energinet.DataHub.MessageArchive.PersistenceModels</RootNamespace>
+  </PropertyGroup>
+
+</Project>
