+++ source/Energinet.DataHub.MessageArchive.Persistence/Services/BlobArchive.cs
@@ -18,11 +18,11 @@ using System.Threading.Tasks;
 using Azure.Storage.Blobs;
 using Azure.Storage.Blobs.Models;
 using Azure.Storage.Blobs.Specialized;
-using Energinet.DataHub.MessageArchive.Domain.Models;
-using Energinet.DataHub.MessageArchive.Domain.Services;
+using Energinet.DataHub.MessageArchive.Processing.Models;
+using Energinet.DataHub.MessageArchive.Processing.Services;
 using Energinet.DataHub.MessageArchive.Utilities;
 
-namespace Energinet.DataHub.MessageArchive.Processing.Services
+namespace Energinet.DataHub.MessageArchive.Persistence.Services
 {
     public class BlobArchive : IBlobArchive
     {
