+++ source/Energinet.DataHub.MessageArchive.Processing/LogParsers/ErrorParsers/XmlErrorParser.cs
@@ -15,7 +15,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Xml.Linq;
-using Energinet.DataHub.MessageArchive.Domain.Models;
+using Energinet.DataHub.MessageArchive.PersistenceModels;
 using Energinet.DataHub.MessageArchive.Utilities;
 
 namespace Energinet.DataHub.MessageArchive.Processing.LogParsers.ErrorParsers
