+++ source/Energinet.DataHub.MessageArchive.Processing/LogParsers/ILogParser.cs
@@ -12,7 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.MessageArchive.Domain.Models;
+using Energinet.DataHub.MessageArchive.PersistenceModels;
+using Energinet.DataHub.MessageArchive.Processing.Models;
 
 namespace Energinet.DataHub.MessageArchive.Processing.LogParsers
 {
