+++ source/Energinet.DataHub.MessageArchive.Processing/LogParsers/LogParserBlobProperties.cs
@@ -13,7 +13,8 @@
 // limitations under the License.
 
 using System.Linq;
-using Energinet.DataHub.MessageArchive.Domain.Models;
+using Energinet.DataHub.MessageArchive.PersistenceModels;
+using Energinet.DataHub.MessageArchive.Processing.Models;
 using Energinet.DataHub.MessageArchive.Utilities;
 
 namespace Energinet.DataHub.MessageArchive.Processing.LogParsers
