+++ source/Energinet.DataHub.MessageArchive.Processing/LogParsers/LogParserErrorResponseXml.cs
@@ -13,8 +13,9 @@
 // limitations under the License.
 
 using System.Xml.Linq;
-using Energinet.DataHub.MessageArchive.Domain.Models;
+using Energinet.DataHub.MessageArchive.PersistenceModels;
 using Energinet.DataHub.MessageArchive.Processing.LogParsers.ErrorParsers;
+using Energinet.DataHub.MessageArchive.Processing.Models;
 using Energinet.DataHub.MessageArchive.Utilities;
 
 namespace Energinet.DataHub.MessageArchive.Processing.LogParsers
