+++ source/Energinet.DataHub.MessageArchive.Tests/Energinet.DataHub.MessageArchive.Tests.csproj
@@ -51,7 +51,7 @@ limitations under the License.
 
     <ItemGroup>
       <ProjectReference Include="..\Energinet.DataHub.MessageArchive.EntryPoint\Energinet.DataHub.MessageArchive.EntryPoint.csproj" />
-      <ProjectReference Include="..\Energinet.DataHub.MessageArchive.Search\Energinet.DataHub.MessageArchive.Search.csproj" />
+      <ProjectReference Include="..\Energinet.DataHub.MessageArchive.Reader\Energinet.DataHub.MessageArchive.Reader.csproj" />
     </ItemGroup>
 
     <ItemGroup>
