+++ source/Energinet.Datahub.MessageArchive.Tests/MockedStorageWriter.cs
@@ -14,8 +14,8 @@
 
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using Energinet.DataHub.MessageArchive.Domain.Repositories;
 using Energinet.DataHub.MessageArchive.Persistence;
+using Energinet.DataHub.MessageArchive.Processing;
 
 namespace Energinet.DataHub.MessageArchive.Tests
 {
